Move ServiceAccount to thunderstore.account
