+++ django/thunderstore/account/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "thunderstore.account.apps.AccountAppConfig"
