+++ django/thunderstore/account/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AccountAppConfig(AppConfig):
+    name = "thunderstore.account"
+    label = "account"
