+++ django/thunderstore/repository/models/__init__.py
@@ -3,5 +3,4 @@ from .package import *
 from .package_download import *
 from .package_rating import *
 from .package_version import *
-from .service_account import *
 from .uploader_identity import *
